Pause simulation when drawing.
Also, fix thread so it doesn't run faster than 100fps. It says its
running at 150fps now... not sure why that is. Before this change, the
work thread was spin-looping if the simulation is not run... I guess
this was putting significant contention on the locks which was causing
the main thread to block for a long time.