Add workaround for String.fromCharCode trick.
I use String.fromCharCode to convert from a typed array buffer to a
string. Chrome seems to store this array on the stack, which exceeds
capacity when the array length is greater than ~120k. Work around for
this is to chunk the array.

Also, remove the brightness/contrast commands when generating
thumbnails.