Fix bug found in pepper.js w.r.t Graphics2D::Flush
The PPAPI documentation says that an "unbound" Graphics2D context will
have its Flush callback called, but a "released" context won't
necessarily.

I haven't seen a case where this fails on Chrome PPAPI, but this fails
consistently in pepper.js. The fix is to keep a reference to the
flushing context, so it is "unbound" but not "released".